refactor(section): name section context modules after their hooks
Both already held a context and its hook in one module. Renaming them
puts every context module in the codebase on one naming pattern.
